ai-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
google-generativeai==0.8.4
pydantic==2.10.6
